requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
psutil==5.9.6
python-dotenv==1.0.0
